Basic Check in - for Universal Binary

+++ Utils/AQGrid/AQGridViewUpdateItem.m
@@ -108,7 +108,7 @@
 			break;
 	}
 	
-	return ( [NSString stringWithFormat: @"%@{index=%u, action=%@, animation=%@, offset=%.02f}", [super description], (unsigned)_index, actionDesc, animationDesc, _offset] );
+	return ( [NSString stringWithFormat: @"%@{index=%u, action=%@, animation=%@, offset=%.02f}", [super description], (unsigned)_index, actionDesc, animationDesc, (float)_offset] );
 }
 
 - (NSComparisonResult) compare: (AQGridViewUpdateItem *) other
